+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Fixed
  * `nativeToScVal` now allows anything to be passed to the `opts.type` specifier. Previously, it was only integer types ([#691](https://github.com/stellar/js-stellar-base/pull/691)).
+ * `Contract.call()` now produces valid `Operation` XDR ([#TODO](https://github.com/stellar/js-stellar-base/pull/TODO)).
 
 
 ## [`v10.0.0-beta.0`](https://github.com/stellar/js-stellar-base/compare/v9.0.0...v10.0.0-beta.0): Protocol 20